+++ RELEASES.md
@@ -21,8 +21,8 @@ and also to avoid C naming collisions.
 - Renamed `TriggerType.LAST` to `LAST_TRADE`
 - Renamed `TriggerType.MARK` to `MARK_PRICE`
 - Renamed `TriggerType.INDEX` to `INDEX_PRICE`
-- Renamed `ComponentState.INITIALIZED` to `POST_INITIALIZED`
-- Moved `manage_gtd_expiry` to `Strategy.submit_order(...)` and Strategy.submit_order_list(...)`
+- Renamed `ComponentState.INITIALIZED` to `READY`
+- Moved `manage_gtd_expiry` to `Strategy.submit_order(...)` and `Strategy.submit_order_list(...)`
 
 ### Enhancements
 - Added `BarSpecification.timedelta` property, thanks @rsmb7z